document primer form dialog in patterns section
