--- lookbook/previews/open_project/dialog_preview/form.html.erb
@@ -1,19 +1,0 @@
-<%= render(Primer::Alpha::Dialog.new(title: dialog_title,
-                                     size: :large,
-                                     id: dialog_id)) do |d| %>
-  <% d.with_show_button { button_text } %>
-  <% d.with_header(variant: :large, mb: 3) %>
-
-  <%= render(Primer::Alpha::Dialog::Body.new) do %>
-    <%= primer_form_with(
-          url: '/abc',
-          id: form_id) do %>
-      Some form elements could be rendered here
-    <% end %>
-  <% end %>
-
-  <%= render(Primer::Alpha::Dialog::Footer.new) do %>
-    <%= render(Primer::ButtonComponent.new(data: { "close-dialog-id": dialog_id })) { I18n.t(:button_cancel) } %>
-    <%= render(Primer::ButtonComponent.new(scheme: :primary, type: :submit, form: form_id)) { I18n.t(:button_apply) } %>
-  <% end %>
-<% end %>
